refactor: move all clusters resource at the same place
